Factor out DiskInterface class into its own source/header files.
This is a TODO in src/ninja_jumble.cc
